Updated train_erac.py

+++ mt/erac/train_erac.py
@@ -160,7 +160,7 @@ def train_erac(src, tgt):
     Q_all = critic(tgt, seq, out_mode=models.LOGIT)
 
     # compute Q(y_{<t}, y_t)
-    Q_mod = Q_all.gather(2, seq[1:].unsqueeze(2)).squeeze(2)
+    Q_mod = Q_all.gather(2, seq[1:].unsqueeze(2).to(torch.int64)).squeeze(2)
 
     # compute V_bar(y_{<t})
     act_log_dist.data.masked_fill_(seq.data[1:].eq(tgt_pad_idx)[:,:,None], 0.)
